fix(cli): align MCP run tools with CLI behavior
