+++ lib/crates/fabro-cli/Cargo.toml
@@ -32,6 +32,7 @@ fabro-install = { path = "../fabro-install" }
 fabro-interview = { path = "../fabro-interview" }
 fabro-mcp = { path = "../fabro-mcp" }
 fabro-mcp-server = { path = "../fabro-mcp-server" }
+fabro-manifest = { path = "../fabro-manifest" }
 fabro-proc = { path = "../fabro-proc" }
 fabro-sandbox = { path = "../fabro-sandbox", features = ["daytona"] }
 fabro-checkpoint = { path = "../fabro-checkpoint" }
