+++ lib/crates/fabro-cli/src/commands/mcp/mod.rs
@@ -4,6 +4,7 @@ use anyhow::{Context as _, Result};
 
 use crate::args::{McpAgent, McpCommand, McpNamespace, ServerConnectionArgs};
 use crate::command_context::CommandContext;
+use crate::user_config;
 
 pub(crate) async fn dispatch(ns: McpNamespace, base_ctx: &CommandContext) -> Result<()> {
     match ns.command {
@@ -26,10 +27,23 @@ fn server_settings(
     base_ctx: &CommandContext,
     connection: &ServerConnectionArgs,
 ) -> Result<fabro_mcp_server::McpServerSettings> {
+    let connection_ctx = base_ctx.with_connection(connection)?;
+    let server_target = user_config::resolve_nondefault_server_target(
+        &connection.target,
+        connection_ctx.user_settings(),
+    )?
+    .map(|target| {
+        target
+            .as_unix_socket_path()
+            .map_or_else(|| target.to_string(), |path| path.display().to_string())
+    });
     Ok(fabro_mcp_server::McpServerSettings {
-        config:   config_settings(connection),
+        config: config_settings(connection),
+        server_target,
+        storage_dir: connection_ctx.storage_dir().to_path_buf(),
+        config_path: connection_ctx.base_config_path().to_path_buf(),
         home_dir: home_dir()?,
-        cwd:      base_ctx.cwd().to_path_buf(),
+        cwd: base_ctx.cwd().to_path_buf(),
     })
 }
